+++ mobile_user_auth/emails.py
@@ -1,4 +1,5 @@
-import random, threading
+import random
+import threading
 
 from django.template.loader import render_to_string
 from django.core.mail import EmailMessage, EmailMultiAlternatives
@@ -6,6 +7,8 @@ from django.conf import settings
 from user import models as account_model
 
 # Email threading
+
+
 class EmailThread(threading.Thread):
     def __init__(self, email):
         self.email = email
@@ -37,6 +40,7 @@ class Util:
             otp.code = code
             otp.save()
 
+
         email_message = EmailMessage(subject=subject, body=message, to=[user.email], from_email=from_email)
         email_message.content_subtype = 'html'
 
@@ -47,23 +51,21 @@ class Util:
         subject = 'Email Verification'
         code = random.randint(1000, 9999)
         message = f'Your email verification OTP is: <strong>{code}</strong>'
-        
+
         otp = account_model.Otp.objects.get_or_none(user=user)
 
         if not otp:
-            account_model.Otp.objects.create(user=user, code=code)    
+            account_model.Otp.objects.create(user=user, code=code)
         else:
             otp.code = code
             otp.save()
-        
+
         email_message = EmailMessage(
-        subject=subject,
-        body=message,
-        to=[user.email],
-        from_email=from_email
-    )
+            subject=subject,
+            body=message,
+            to=[user.email],
+            from_email=from_email
+        )
         email_message.content_subtype = 'html'
 
         EmailThread(email_message).start()
-            
-
