+++ user/admin.py
@@ -1,8 +1,9 @@
 from django.contrib import admin
 
-from user.models import User, EntryCode
+from user.models import Otp, User, EntryCode
 
 # Register your models here.
 
 admin.site.register(User)
 admin.site.register(EntryCode)
+admin.site.register(Otp)
