+++ user/urls.py
@@ -1,6 +1,6 @@
 from rest_framework.routers import DefaultRouter
 from django.urls import path, include
-from .views import DashboardViewSet, LoginViewSet, UsersViewSet, ForgotPasswordView, ResetPasswordView
+from .views import DashboardViewSet, LoginViewSet, UsersViewSet
 
 
 router = DefaultRouter()
@@ -11,7 +11,5 @@ router.register(r'users', UsersViewSet, basename="users")
 
 urlpatterns = [
     path('', include(router.urls)),
-    path('forgot-password/', ForgotPasswordView.as_view(), name='forgot-password'),
-    path('reset-password/<str:uidb64>/<str:token>/',
-         ResetPasswordView.as_view(), name='reset-user-pass'),
+
 ]
